import restructuring, begin of test generalization

+++ src/spflow/torch/sampling/general/layers/leaves/parametric/gaussian.py
@@ -90,7 +90,7 @@ def sample(
         sampling_mask = marg_ids & instance_ids_mask.bool().to(layer.mean.device)
         sampling_ids = torch.where(sampling_mask)[0]
 
-        data[torch.meshgrid(sampling_ids, torch.tensor(node_scope.query), indexing="ij")] = (
+        data[torch.meshgrid(sampling_ids, torch.tensor(node_scope.query, dtype=torch.int64), indexing="ij")] = (
             layer.dist([node_id]).sample((sampling_mask.sum(),)).to(layer.mean.device)
         )
